requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.14.0
django-filter==22.1
django-jasmine==0.5.1
djangorestframework==3.14.0
filelock==3.8.0
platformdirs==2.5.4
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.16.7
